tracing/osnoise: Fix duration type
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2174944

commit 022632f6c43a86f2135642dccd5686de318e861d
Author: Daniel Bristot de Oliveira <bristot@kernel.org>
Date:   Thu Nov 17 14:46:17 2022 +0100

	tracing/osnoise: Fix duration type

	The duration type is a 64 long value, not an int. This was
	causing some long noise to report wrong values.

	Change the duration to a 64 bits value.

	Link: https://lkml.kernel.org/r/a93d8a8378c7973e9c609de05826533c9e977939.1668692096.git.bristot@kernel.org

	Cc: stable@vger.kernel.org
	Cc: Daniel Bristot de Oliveira <bristot@kernel.org>
	Cc: Steven Rostedt <rostedt@goodmis.org>
	Cc: Masami Hiramatsu <mhiramat@kernel.org>
	Cc: Jonathan Corbet <corbet@lwn.net>
	Fixes: bce29ac9ce0b ("trace: Add osnoise tracer")
	Acked-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>

+++ kernel/trace/trace_osnoise.c
@@ -918,7 +918,7 @@ void osnoise_trace_irq_entry(int id)
 void osnoise_trace_irq_exit(int id, const char *desc)
 {
 	struct osnoise_variables *osn_var = this_cpu_osn_var();
-	int duration;
+	s64 duration;
 
 	if (!osn_var->sampling)
 		return;
@@ -1049,7 +1049,7 @@ static void trace_softirq_entry_callback(void *data, unsigned int vec_nr)
 static void trace_softirq_exit_callback(void *data, unsigned int vec_nr)
 {
 	struct osnoise_variables *osn_var = this_cpu_osn_var();
-	int duration;
+	s64 duration;
 
 	if (!osn_var->sampling)
 		return;
@@ -1145,7 +1145,7 @@ thread_entry(struct osnoise_variables *osn_var, struct task_struct *t)
 static void
 thread_exit(struct osnoise_variables *osn_var, struct task_struct *t)
 {
-	int duration;
+	s64 duration;
 
 	if (!osn_var->sampling)
 		return;
